Refactored LawProcessor and Created LawConstructor
